requirements.txt
tensorflow==2.13.0
streamlit==1.28.0
opencv-python==4.8.1.78
pillow==10.1.0
matplotlib==3.7.1
seaborn==0.13.0
pandas==2.1.4
numpy==1.24.3
pathlib==1.0.1